Move developer spotlight newsletter signup to bottom of home page
Extract NewsletterSignup component, render it after the developer directory
and before the footer on Welcome. Remove newsletter block from Hero so other
Hero pages stay unchanged. Extend WelcomePageTourAnchorsTest for the new layout.

Made-with: Cursor
